Add variable-qty, size pricing, currency via iso4217parse, merge Inventory.xlsx

+++ packages/shopyo_ecommerce/shopyo_ecommerce/pos/models.py
@@ -42,7 +42,7 @@ class TransactionItem(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     transaction_id = db.Column(db.Integer, db.ForeignKey('shopyo_ecommerce_transactions.id'), nullable=False)
     product_barcode = db.Column(db.String(100), nullable=False)
-    quantity = db.Column(db.Integer, nullable=False)
+    quantity = db.Column(db.Numeric(10, 2), nullable=False)
     unit_price = db.Column(db.Numeric(10, 2), nullable=False)
 
     def add(self):
